Adapt client to the server
Refactor HarvestJournalEntry into a public class HarvestEntry.
The idea is to use this HarvestEntry to handle the interaction
between all classes.

Modify serialization format to adapt it to the back-end, the idea
is to save redundant information while transmitting the report.
This is achieved by indexing the apps usage sesions by package
name.

Rename SQLite DB to match server naming conventions, and finally,
do a generate code cleanup.

Signed-off-by: Martin Abente Lahaye <tch@sugarlabs.org>