Refactor the mechanism to run the service regularly
Move all tracking logic from HarvestService to a new HarvestWatcher
class. This new class will also handle the run-regularly logic. To
do this, the new class implements the Runnable interface and uses
the Handler class.

Remove all AlarmManager related code, as explained adove, this is
replace by the new class. This is change is due the fact that this
Alarm approach, did not let the device go sleep.

Improve logging messages specifying which class it belongs to, to
improve debugging clarity.

Signed-off-by: Martin Abente Lahaye <tch@sugarlabs.org>