Monitor network traffic usage
Accumulates received and transmitted bytes and aggregates
these values for each day.

Implements an alternative for missing Android's TrafficStats,
called HarvestTrafficStats. The methods provided by this
implementation are similar to Android's, except that this is
not a singleton and must initialized for each reading.

Implements a journal for traffic data. Similarly to usage time
journal, it works a in-memory buffer for traffic stats.

Implements a store for persisting data in a SQLite database.

Modifies HarvestReporter to include traffic stats entries in
the report.

Add new settings in HarvestSetting, to control how often it will
read traffic stats and how often it will persist it.

Signed-off-by: Martin Abente Lahaye <tch@sugarlabs.org>